Had some timer problems, not sure what they all were, but part of it was me trying to clean up the timer process as part of the debug reset/reload behavior in flask.  Turns out that I do not understand that as well as I thought, I guess.
Anyhow, this is my current attempt at making the timer live on.  A worth noting thought on this one is that the keep alive behavior is provoked by the bike observer, and not the timer, which should be a moot point, but that could come up later somehow.  Have to wait and see.