Server side code now uses environmental variables to locate data files.
Additionally, I found and addressed a strange problem in the Arduino RPM monitor that sporadically showed readings above 60,000 RPMs.  An impossible speed for every reason.  My solution was to show readings above a certain value as missing.  Might be ugly, but with that data missing, I can think of no other way to handle the situation.