Everything is wrong with this.  I tried for a very long time and made no progress at all.
When an exception is raised in another thread sys.excepthook does not seem to catch it.

UPDATE: So, in typing this out, I realized that I had been focused on socketio.  Because the problem always seemed to involve sockets.  I decided to look into threading and sys.excepthook and found a solution, or atleast a workaround that will meet my needs.

I am currently running Python 3.6.3.  I really thought that I had updated atleast once since I got it, but appearantly not.  This is my original python install.  SO says the problem was fixed in Python 3.8, which is what I figured I had.

That is the real fix.  Update Python.

Everything seems to be logging errors as expected now.  The error itself can now be looked into.