So, I have no link to the reports page, and report filing is still a bit hackish, but I wrote a selenium script which does nice data cluster detection and I used that to fill in the missing days in the report view.
Also, the tabata timer is not editable, even though, it really seems like it should be.

Reports can be deleted, or atleast retired, from the /info url.

Comments can be created and edited using a function on the info page.  The edit view logic, well, it is not.

Finally, clicking on a report brings up the graph of relevant data.  D3 graphing page/endpoint can now display any historical period, not just the latest recording.