Minor code cleanup in graph.js.
Added display of times as a prettified minutes:seconds format to schedule editor.  For the moment, both total seconds and mm:ss are displayed.  This makes the feeling cluttered, I would prefer one or the other, but have not figured out an intuitive way to do that.

Scheduling can now be modified through the online editor and multiple schedules can be recoreded at one time.  The three buttons for start a new schedule, save updates to an existing schedule and save updates of existing schedule as a new schedule are, well, I would preffer to have one or maybe two.  Three seems like too much there, also.

Updateing a schedule that has been deleted from the server results in an unhelpful error, that I have not yet figured out how to address, and opening a schedule in the editor, modifying it somehow, and then clicking on another schedule from the list will load the remote schedule into the editor.  This destroys the modifications and so far, no warning is displayed, the changes just go away.  I have not figured out how to mark a schedule as modified, or how to handle that situation, but something should be said first.

Finally, this whole thing has now become more complicated than I had expected.  Schedule editing should really have its own page, but that should be a thing that happens after vuex and probably the conversion of everything to a single vue instance.

Vuex and vue-router should be in this app's future, with maybe some client side tests to boot.